+++ x/epochs/abci.go
@@ -15,28 +15,18 @@ import (
 func BeginBlocker(ctx sdk.Context, k keeper.Keeper) {
 	defer telemetry.ModuleMeasureSince(types.ModuleName, time.Now(), telemetry.MetricKeyBeginBlocker)
 	k.IterateEpochInfo(ctx, func(index int64, epochInfo types.EpochInfo) (stop bool) {
-		// If blocktimme < initial epoch start time, return
 		if ctx.BlockTime().Before(epochInfo.StartTime) {
 			return false
 		}
 
-		logger := k.Logger(ctx)
-
-		// Epoch has not started yet
-		shouldInitialEpochStart := !epochInfo.EpochCountingStarted
-
-		epochEndTime := epochInfo.CurrentEpochStartTime.Add(epochInfo.Duration)
-		// StartTime is set to a pinpointed timestamp that is after the default value of CurrentEpochStartTime (=0).
-		shouldEpochStart := shouldInitialEpochStart || ctx.BlockTime().After(epochEndTime)
-
-		if !shouldEpochStart {
+		if !shouldEpochStart(epochInfo, ctx) {
 			return false
 		}
 
-		// we deduced that a new epoch tick should happen
 		epochInfo.CurrentEpochStartHeight = ctx.BlockHeight()
 
-		if shouldInitialEpochStart {
+		logger := k.Logger(ctx)
+		if !epochInfo.EpochCountingStarted {
 			epochInfo.EpochCountingStarted = true
 			epochInfo.CurrentEpoch = 1
 			epochInfo.CurrentEpochStartTime = epochInfo.StartTime
@@ -62,9 +52,24 @@ func BeginBlocker(ctx sdk.Context, k keeper.Keeper) {
 				sdk.NewAttribute(types.AttributeEpochStartTime, fmt.Sprintf("%d", epochInfo.CurrentEpochStartTime.Unix())),
 			),
 		)
-		k.SetEpochInfo(ctx, epochInfo)
+		k.UpsertEpochInfo(ctx, epochInfo)
 		k.BeforeEpochStart(ctx, epochInfo.Identifier, epochInfo.CurrentEpoch)
 
 		return false
 	})
+}
+
+// shouldEpochStart checks if the epoch should start.
+// an epoch is ready to start if:
+// - it has not yet been initialized.
+// - the current epoch end time is before the current block time
+func shouldEpochStart(epochInfo types.EpochInfo, ctx sdk.Context) bool {
+	// Epoch has not started yet
+	if !epochInfo.EpochCountingStarted {
+		return true
+	}
+
+	epochEndTime := epochInfo.CurrentEpochStartTime.Add(epochInfo.Duration)
+
+	return ctx.BlockTime().After(epochEndTime)
 }