+++ x/epochs/genesis.go
@@ -9,7 +9,6 @@ import (
 
 // InitGenesis sets epoch info from genesis
 func InitGenesis(ctx sdk.Context, k keeper.Keeper, genState types.GenesisState) {
-	// set epoch info from genesis
 	for _, epoch := range genState.Epochs {
 		if err := k.AddEpochInfo(ctx, epoch); err != nil {
 			panic(err)
@@ -21,5 +20,6 @@ func InitGenesis(ctx sdk.Context, k keeper.Keeper, genState types.GenesisState) 
 func ExportGenesis(ctx sdk.Context, k keeper.Keeper) *types.GenesisState {
 	genesis := types.DefaultGenesis()
 	genesis.Epochs = k.AllEpochInfos(ctx)
+
 	return genesis
 }