+++ x/epochs/genesis_test.go
@@ -4,10 +4,9 @@ import (
 	"testing"
 	"time"
 
-	"github.com/NibiruChain/nibiru/simapp"
-
 	"github.com/stretchr/testify/require"
 
+	"github.com/NibiruChain/nibiru/simapp"
 	"github.com/NibiruChain/nibiru/x/epochs"
 	"github.com/NibiruChain/nibiru/x/epochs/types"
 )
@@ -24,7 +23,7 @@ func TestEpochsExportGenesis(t *testing.T) {
 	require.Equal(t, genesis.Epochs[0].Identifier, "15 min")
 	require.Equal(t, genesis.Epochs[0].StartTime, chainStartTime)
 	require.Equal(t, genesis.Epochs[0].Duration, time.Minute*15)
-	require.Equal(t, genesis.Epochs[0].CurrentEpoch, int64(0))
+	require.Equal(t, genesis.Epochs[0].CurrentEpoch, uint64(0))
 	require.Equal(t, genesis.Epochs[0].CurrentEpochStartHeight, chainStartHeight)
 	require.Equal(t, genesis.Epochs[0].CurrentEpochStartTime, chainStartTime)
 	require.Equal(t, genesis.Epochs[0].EpochCountingStarted, false)
@@ -32,7 +31,7 @@ func TestEpochsExportGenesis(t *testing.T) {
 	require.Equal(t, genesis.Epochs[1].Identifier, "30 min")
 	require.Equal(t, genesis.Epochs[1].StartTime, chainStartTime)
 	require.Equal(t, genesis.Epochs[1].Duration, time.Minute*30)
-	require.Equal(t, genesis.Epochs[1].CurrentEpoch, int64(0))
+	require.Equal(t, genesis.Epochs[1].CurrentEpoch, uint64(0))
 	require.Equal(t, genesis.Epochs[1].CurrentEpochStartHeight, chainStartHeight)
 	require.Equal(t, genesis.Epochs[1].CurrentEpochStartTime, chainStartTime)
 	require.Equal(t, genesis.Epochs[1].EpochCountingStarted, false)
@@ -40,7 +39,7 @@ func TestEpochsExportGenesis(t *testing.T) {
 	require.Equal(t, genesis.Epochs[2].Identifier, "day")
 	require.Equal(t, genesis.Epochs[2].StartTime, chainStartTime)
 	require.Equal(t, genesis.Epochs[2].Duration, time.Hour*24)
-	require.Equal(t, genesis.Epochs[2].CurrentEpoch, int64(0))
+	require.Equal(t, genesis.Epochs[2].CurrentEpoch, uint64(0))
 	require.Equal(t, genesis.Epochs[2].CurrentEpochStartHeight, chainStartHeight)
 	require.Equal(t, genesis.Epochs[2].CurrentEpochStartTime, chainStartTime)
 	require.Equal(t, genesis.Epochs[2].EpochCountingStarted, false)
@@ -48,7 +47,7 @@ func TestEpochsExportGenesis(t *testing.T) {
 	require.Equal(t, genesis.Epochs[3].Identifier, "week")
 	require.Equal(t, genesis.Epochs[3].StartTime, chainStartTime)
 	require.Equal(t, genesis.Epochs[3].Duration, time.Hour*24*7)
-	require.Equal(t, genesis.Epochs[3].CurrentEpoch, int64(0))
+	require.Equal(t, genesis.Epochs[3].CurrentEpoch, uint64(0))
 	require.Equal(t, genesis.Epochs[3].CurrentEpochStartHeight, chainStartHeight)
 	require.Equal(t, genesis.Epochs[3].CurrentEpochStartTime, chainStartTime)
 	require.Equal(t, genesis.Epochs[3].EpochCountingStarted, false)
@@ -111,7 +110,7 @@ func TestEpochsInitGenesis(t *testing.T) {
 	require.Equal(t, epochInfo.Identifier, "monthly")
 	require.Equal(t, epochInfo.StartTime.UTC().String(), now.UTC().String())
 	require.Equal(t, epochInfo.Duration, time.Hour*24)
-	require.Equal(t, epochInfo.CurrentEpoch, int64(0))
+	require.Equal(t, epochInfo.CurrentEpoch, uint64(0))
 	require.Equal(t, epochInfo.CurrentEpochStartHeight, ctx.BlockHeight())
 	require.Equal(t, epochInfo.CurrentEpochStartTime.UTC().String(), time.Time{}.String())
 	require.Equal(t, epochInfo.EpochCountingStarted, true)