+++ x/epochs/keeper/epoch.go
@@ -25,6 +25,14 @@ func (k Keeper) GetEpochInfo(ctx sdk.Context, identifier string) types.EpochInfo
 	return epoch
 }
 
+// EpochExists checks if the epoch exists
+func (k Keeper) EpochExists(ctx sdk.Context, identifier string) bool {
+	store := ctx.KVStore(k.storeKey)
+	b := store.Get(append(types.KeyPrefixEpoch, []byte(identifier)...))
+
+	return b != nil
+}
+
 // AddEpochInfo adds a new epoch info. Will return an error if the epoch fails validation,
 // or re-uses an existing identifier.
 // This method also sets the start time if left unset, and sets the epoch start height.
@@ -32,8 +40,8 @@ func (k Keeper) AddEpochInfo(ctx sdk.Context, epoch types.EpochInfo) error {
 	if err := epoch.Validate(); err != nil {
 		return err
 	}
-	// Check if identifier already exists
-	if (k.GetEpochInfo(ctx, epoch.Identifier) != types.EpochInfo{}) {
+
+	if k.EpochExists(ctx, epoch.Identifier) {
 		return fmt.Errorf("epoch with identifier %s already exists", epoch.Identifier)
 	}
 
@@ -41,13 +49,16 @@ func (k Keeper) AddEpochInfo(ctx sdk.Context, epoch types.EpochInfo) error {
 	if epoch.StartTime.Equal(time.Time{}) {
 		epoch.StartTime = ctx.BlockTime()
 	}
+
 	epoch.CurrentEpochStartHeight = ctx.BlockHeight()
-	k.SetEpochInfo(ctx, epoch)
+
+	k.UpsertEpochInfo(ctx, epoch)
+
 	return nil
 }
 
-// setEpochInfo set epoch info.
-func (k Keeper) SetEpochInfo(ctx sdk.Context, epoch types.EpochInfo) {
+// UpsertEpochInfo inserts the epoch if does not exist, and overwrites it if it does.
+func (k Keeper) UpsertEpochInfo(ctx sdk.Context, epoch types.EpochInfo) {
 	store := ctx.KVStore(k.storeKey)
 	value, err := proto.Marshal(&epoch)
 	if err != nil {
@@ -94,16 +105,4 @@ func (k Keeper) AllEpochInfos(ctx sdk.Context) []types.EpochInfo {
 		return false
 	})
 	return epochs
-}
-
-// NumBlocksSinceEpochStart returns the number of blocks since the epoch started.
-// if the epoch started on block N, then calling this during block N (after BeforeEpochStart)
-// would return 0.
-// Calling it any point in block N+1 (assuming the epoch doesn't increment) would return 1.
-func (k Keeper) NumBlocksSinceEpochStart(ctx sdk.Context, identifier string) (int64, error) {
-	epoch := k.GetEpochInfo(ctx, identifier)
-	if (epoch == types.EpochInfo{}) {
-		return 0, fmt.Errorf("epoch with identifier %s not found", identifier)
-	}
-	return ctx.BlockHeight() - epoch.CurrentEpochStartHeight, nil
 }