+++ x/epochs/keeper/hooks.go
@@ -4,10 +4,12 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 )
 
-func (k Keeper) AfterEpochEnd(ctx sdk.Context, identifier string, epochNumber int64) {
+// AfterEpochEnd epoch hook
+func (k Keeper) AfterEpochEnd(ctx sdk.Context, identifier string, epochNumber uint64) {
 	k.hooks.AfterEpochEnd(ctx, identifier, epochNumber)
 }
 
-func (k Keeper) BeforeEpochStart(ctx sdk.Context, identifier string, epochNumber int64) {
+// BeforeEpochStart epoch hook
+func (k Keeper) BeforeEpochStart(ctx sdk.Context, identifier string, epochNumber uint64) {
 	k.hooks.BeforeEpochStart(ctx, identifier, epochNumber)
 }