+++ x/perp/keeper/hooks.go
@@ -12,10 +12,10 @@ import (
 	"github.com/NibiruChain/nibiru/x/perp/types"
 )
 
-func (k Keeper) BeforeEpochStart(ctx sdk.Context, epochIdentifier string, epochNumber int64) {
+func (k Keeper) BeforeEpochStart(ctx sdk.Context, epochIdentifier string, epochNumber uint64) {
 }
 
-func (k Keeper) AfterEpochEnd(ctx sdk.Context, epochIdentifier string, _ int64) {
+func (k Keeper) AfterEpochEnd(ctx sdk.Context, epochIdentifier string, _ uint64) {
 	params := k.GetParams(ctx)
 	if epochIdentifier != params.FundingRateInterval || params.Stopped {
 		return
@@ -84,16 +84,16 @@ type Hooks struct {
 
 var _ epochstypes.EpochHooks = Hooks{}
 
-// Return the wrapper struct.
+// Hooks Return the wrapper struct.
 func (k Keeper) Hooks() Hooks {
 	return Hooks{k}
 }
 
-// epochs hooks.
-func (h Hooks) BeforeEpochStart(ctx sdk.Context, epochIdentifier string, epochNumber int64) {
+// BeforeEpochStart epochs hooks.
+func (h Hooks) BeforeEpochStart(ctx sdk.Context, epochIdentifier string, epochNumber uint64) {
 	h.k.BeforeEpochStart(ctx, epochIdentifier, epochNumber)
 }
 
-func (h Hooks) AfterEpochEnd(ctx sdk.Context, epochIdentifier string, epochNumber int64) {
+func (h Hooks) AfterEpochEnd(ctx sdk.Context, epochIdentifier string, epochNumber uint64) {
 	h.k.AfterEpochEnd(ctx, epochIdentifier, epochNumber)
 }